// src/shared/lib/models/DatePeriodFilterType.ts
export enum DatePeriodFilterType {
  ALL = 'all',
  TODAY = 'today',
  YESTERDAY = 'yesterday',
  CURRENT_WEEK = 'current_week',
  LAST_WEEK = 'last_week',
  CURRENT_MONTH = 'current_month',
  LAST_MONTH = 'last_month',
  CURRENT_QUARTER = 'current_quarter',
  LAST_QUARTER = 'last_quarter',
  PERIOD = 'period',
}